Notebook capturing some LSTM Tuning Experiments
Experiment1: Using data without TA indicators as features
Result: 66%  Accuracy

Experiment 2: Using TA indicators as features
Result: 83% Accuracy

Discover an Error - shifting the time series the wrong way - hence the predictions were laggy. It should be shifted +1.. not -1 
Now the results are not laggy

Suspect that accuracy is low because of the scaling based normatilization done for the entire time series of data. The price increase over time may be causing issues.

Need to develop a new normalization function that changes over time - based on rate of change in stock price